All IDs set to 8 bits

+++ fsm_rgb_light.c
@@ -24,7 +24,7 @@
 /* Other auxiliary functions */
 
 /* Public functions -----------------------------------------------------------*/
-fsm_rgb_light_t *fsm_rgb_light_new(uint32_t rgb_light_id)
+fsm_rgb_light_t *fsm_rgb_light_new(uint8_t rgb_light_id)
 {
     fsm_rgb_light_t *p_fsm_rgb_light = malloc(sizeof(fsm_rgb_light_t)); /* Do malloc to reserve memory of all other FSM elements, although it is interpreted as fsm_t (the first element of the structure) */
     fsm_rgb_light_init(p_fsm_rgb_light, rgb_light_id); /* Initialize the FSM */
